Bump nuget packages, align role names

+++ src/backend/api/Fusion.Resources.Domain/Commands/Departments/DeleteDelegatedResourceOwner.cs
@@ -3,15 +3,13 @@ using Fusion.Integration.Roles;
 using Fusion.Resources.Database;
 using MediatR;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Fusion.Resources.Domain.Commands.Departments
 {
-    public  class DeleteDelegatedResourceOwner : IRequest<bool>
+    public class DeleteDelegatedResourceOwner : IRequest<bool>
     {
         private readonly string departmentId;
         private readonly Guid delegatedOwnerAzureUniqueId;
@@ -24,21 +22,33 @@ namespace Fusion.Resources.Domain.Commands.Departments
 
         public class Handler : IRequestHandler<DeleteDelegatedResourceOwner, bool>
         {
+            private readonly ResourcesDbContext db;
             private readonly IFusionRolesClient rolesClient;
 
-            public Handler(IFusionRolesClient rolesClient)
+            public Handler(ResourcesDbContext db, IFusionRolesClient rolesClient)
             {
+                this.db = db;
                 this.rolesClient = rolesClient;
             }
 
             public async Task<bool> Handle(DeleteDelegatedResourceOwner request, CancellationToken cancellationToken)
             {
+                var query = db.DepartmentResponsibles
+                    .Where(x => x.DepartmentId == request.departmentId
+                                && x.ResponsibleAzureObjectId == request.delegatedOwnerAzureUniqueId);
+                db.DepartmentResponsibles.RemoveRange(query);
+
+                var wasDeleted = await db.SaveChangesAsync(cancellationToken) > 0;
+
+                if (!wasDeleted)
+                    return false;
+
                 var deleted = await rolesClient.DeleteRolesAsync(
                     new PersonIdentifier(request.delegatedOwnerAzureUniqueId),
-                    q => q.WhereRoleName(Roles.ResourceOwner).WhereScopeValue(request.departmentId)
+                    q => q.WhereRoleName(AccessRoles.ResourceOwner).WhereScopeValue(request.departmentId)
                 );
 
-                return deleted.Count() > 0;
+                return deleted.Any();
             }
         }
     }
